Removed Orders model for now.

+++ db/migrate/20201110225255_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[6.0]
   def change
     create_table :products do |t|
       t.string :name, null: false
-      t.integer :price, null: false
+      t.decimal :price, null: false
       t.bigint :user_id, null: false
       t.timestamps
     end
